backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
sqlalchemy
psycopg2-binary
pgvector
langchain
langchain-openai
langchain-postgres
langchain-community
langchain-text-splitters
langchain-aws
openai
pydantic
pydantic-settings
yt-dlp
pypdf
python-docx
fpdf
boto3
python-pptx
